Intial Commit
Topics with Learning Tools materials sort of work (the topic is created in puppeteer, then the UI is turned over to the user to drag in a Learning Tool widget and choose the right tool, then the remainder of the tool config is scripted), then the user clicks back to Topics, and the next topic is created.

Basically: Puppeteer’s drag-and-drop scripting and Blackbaud’s UI seem not to be immediately compatible.

I started down the path of creating assignments, but the overhaul of the assignments UI using SKY UI renders it very vexing to scrape. I tried keyboard controls only, but I think there are some significant timing issues.

The next step is to revisit just sending payloads directly to the front-end UI, which would (alas) probably be more reliable.